+++ core/src/main/scala/de/sciss/fscape/graph/RepeatWindow.scala
@@ -27,6 +27,6 @@ final case class RepeatWindow(in: GE, size: GE = 1, num: GE = 2) extends UGenSou
 
   private[fscape] def makeStream(args: Vec[StreamIn])(implicit b: stream.Builder): StreamOut = {
     val Vec(in, size, num) = args
-    stream.RepeatWindow(in = in.toDouble, size = size.toInt, num = num.toInt)
+    stream.RepeatWindow(in = in.toDouble, size = size.toInt, num = num.toLong)
   }
 }
